@@ -29,7 +29,7 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.10,<3.13"
+python = "~3.12"
 aiofiles = "^0.7.0"
 aiohttp = "^3.8.1"
 biopython = "^1.81"
@@ -38,7 +38,7 @@ orjson = "^3.9.9"
 pydantic-factories = "^1.17.3"
 pyfixtures = "^1.0.0"
 sentry-sdk = "^2.3.1"
-virtool-core = "^12.0.0"
+virtool-core = "^14.0.0"
 
 [tool.poetry.scripts]
 run-workflow = "virtool_workflow.cli:cli_main"
@@ -71,8 +71,14 @@ exclude = [
     ".ruff_cache",
     "__pypackages__",
 ]
+target-version = "py312"
 
 [tool.ruff.lint]
+ignore = [
+    "ANN101",
+    "D203",
+    "D213"
+]
 select = ["ALL"]
 
 [build-system]
