@@ -1,61 +1,62 @@
+"""Custom exceptions for ``virtool_workflow``."""
+
 from subprocess import SubprocessError
 
 
-class IllegalJobArguments(ValueError):
-    """The `job.args` dict is in an illegal state."""
+class JobAlreadyAcquiredError(Exception):
+    """Raised when an attempt is made to reacquire a job."""
 
-
-
-class InsufficientJobRights(Exception):
-    ...
-
-
-class JobAlreadyAcquired(Exception):
-    def __init__(self, job_id: str):
-        super(JobAlreadyAcquired, self).__init__(
+    def __init__(self, job_id: str) -> None:
+        """Initialize the exception with a message containing the job ID."""
+        super().__init__(
             f"Job {job_id} is has already been acquired.",
         )
-
-
-class JobAlreadyFinalized(Exception):
-    ...
 
 
 class JobsAPIError(Exception):
     """A base exception for errors due to HTTP errors from the jobs API."""
 
 
-
-class JobsAPIBadRequest(JobsAPIError):
-    """A ``400 Bad Request`` response from the jobs API."""
+class JobsAPIBadRequestError(JobsAPIError):
+    """A ``400 Bad Request`` response was received from the jobs API."""
 
     status = 400
 
 
-class JobsAPIForbidden(JobsAPIError):
+class JobsAPIForbiddenError(JobsAPIError):
+    """A ``403 Forbidden`` response was received from the jobs API."""
+
     status = 403
 
 
-class JobsAPINotFound(JobsAPIError):
+class JobsAPINotFoundError(JobsAPIError):
+    """A ``404 Not Found`` response was received from the jobs API."""
+
     status = 404
 
 
-class JobsAPIConflict(JobsAPIError):
+class JobsAPIConflictError(JobsAPIError):
+    """A ``409 Conflict`` response was received from the jobs API."""
+
     status = 409
 
 
 class JobsAPIServerError(JobsAPIError):
+    """A ``500 Internal Server Error`` response was received from the jobs API."""
+
     status = 500
 
 
-class MissingJobArgument(ValueError):
+class MissingJobArgumentError(ValueError):
     """The `job.args` dict is missing a required key for some funcionality."""
 
 
+class WorkflowStepsError(Exception):
+    """Raised when no workflow steps are found in a module."""
 
-class NotFound(KeyError):
-    ...
+    def __init__(self, module: str) -> None:
+        super().__init__(f"No workflow steps could be found in {module}")
 
 
-class SubprocessFailed(SubprocessError):
+class SubprocessFailedError(SubprocessError):
     """Subprocess exited with non-zero status during a workflow."""
